code adapted to vs2019 implemented multithreaded preloading of meshes

+++ Code/Engine/Base/Utils/LerpAnimator1D.cpp
@@ -45,7 +45,7 @@ bool CLerpAnimator1D::Update (float deltaTime, float &value)
 		break;
 	case FUNC_DECREMENT:
 		{
-			mu = sqrt(mu);
+			mu = (float)sqrt(mu);
 		}
 		break;
 	}
