+++ Code/Engine/Graphics/Camera/ThPSCamera.cpp
@@ -36,9 +36,9 @@ Vect3f CThPSCamera::GetEye () const
 	Vect3f pos	= m_pObject3D->GetPosition();
 
 	//Pasamos de coordenadas esfericas a coordenadas cartesianas
-	Vect3f vEyePt(	m_fZoom * cos(yaw) * cos(pitch), 
-		m_fZoom * sin(pitch),
-		m_fZoom * sin(yaw) * cos(pitch) );
+	Vect3f vEyePt((float)(m_fZoom * cos(yaw) * cos(pitch)),
+		(float)(m_fZoom * sin(pitch)),
+		(float)(m_fZoom * sin(yaw) * cos(pitch)) );
 
 	return (pos - vEyePt); 
 }
@@ -50,9 +50,9 @@ Vect3f CThPSCamera::GetVecUp () const
 	float yaw		= m_pObject3D->GetYaw();
 	float pitch	= m_pObject3D->GetPitch();
 
-	Vect3f vUpVec(	-cos(yaw) * sin(pitch), 
-		cos(pitch), 
-		-sin(yaw) * sin(pitch) );
+	Vect3f vUpVec((float)(-cos(yaw) * sin(pitch)),
+		(float)cos(pitch),
+		(float)(-sin(yaw) * sin(pitch)) );
 
 	return vUpVec;
 }
