+++ Code/Engine/Graphics/Object3D.cpp
@@ -90,9 +90,9 @@ CPhysicUserData* CObject3D::GetPhysicUserData() const
 Vect3f CObject3D::xzFromAngle( float radians ) 
 {
 	Vect3f l_aux;
-	l_aux.x = -sin( radians );
+	l_aux.x = (float)-sin( radians );
 	l_aux.y = float( 0.0f );
-	l_aux.z = cos( radians );
+	l_aux.z = (float)cos( radians );
 	return l_aux;
 }
