+++ Code/Engine/Graphics/TorusEmitter.cpp
@@ -47,8 +47,8 @@ void CTorusEmitter::GenerateParticlePosition(CParticle * particle)
 Vect3f CTorusEmitter::xzFromAngle( float radians ) 
 {
 	Vect3f l_aux;
-	l_aux.x = sin( radians );
+	l_aux.x = (float)sin( radians );
 	l_aux.y = float( 0.0f );
-	l_aux.z = -cos( radians );
+	l_aux.z = (float)(-cos( radians ));
 	return l_aux;
 }
